Commit the new version of netmap-fwd.
This version supports VLANs and also connects the host ring with the NIC.

With the host ring bridged to NIC, the system network behaves the same as
before.

Use the -H flag to not open the host ring and answer the ICMP and ARP packets
in netmap-fwd (instead of bridge them to the host).

Many bug fixes.